Move MainPage to ui.main package
